+++ sdk/appconfiguration/Azure.Data.AppConfiguration/tests/AppConfigurationTestEnvironment.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
+using Azure.Core;
 using Azure.Core.TestFramework;
 using Azure.Identity;
 
@@ -12,6 +13,9 @@ namespace Azure.Data.AppConfiguration
         public string ConnectionString => GetRecordedVariable("APPCONFIGURATION_CONNECTION_STRING", options => options.HasSecretConnectionStringParameter("secret", SanitizedValue.Base64));
         public string Endpoint => GetRecordedVariable("APPCONFIGURATION_ENDPOINT_STRING");
         public string SecretId => GetRecordedVariable("KEYVAULT_SECRET_URL");
+
+        protected override TokenCredential CreateDeveloperCredential()
+            => new AzureCliCredential();
 
         public AppConfigurationAudience GetAudience()
         {